Tied up loose ends - added form validation in Surajs Map Screen, Aligned text in Suraj's confirm screen, added font to some text & corrected the pixel overflow problem
