+++ front-end/rideshare/lib/screens/add_ride_map_screen.dart
@@ -5,7 +5,7 @@ import 'package:google_maps_flutter/google_maps_flutter.dart';
 import 'package:flutter_dotenv/flutter_dotenv.dart';
 import 'package:flutter_google_places/flutter_google_places.dart';
 import 'package:google_maps_webservice/places.dart';
-import 'submit_ride_confirm_screen.dart';
+import 'add_ride_confirm_screen.dart';
 
 class AddRideMapScreen extends StatefulWidget {
   final String date;
@@ -26,29 +26,9 @@ class _AddRideMapScreenState extends State<AddRideMapScreen> {
   final _startsearchFieldController = TextEditingController();
   final _endsearchFieldController = TextEditingController();
   final gmaps_api_key = dotenv.env["GOOGLE_MAPS_API_KEY"] ?? "";
+  final GlobalKey<FormState> _formKey = GlobalKey<FormState>();
   LatLng? _sourceLatLng;
   LatLng? _destinationLatLng;
-
-  // In _AddRideMapScreenState class
-  // In _AddRideMapScreenState class
-  void _goToConfirmationScreen(BuildContext context) {
-    if (_startsearchFieldController.text.isNotEmpty && _endsearchFieldController.text.isNotEmpty && _sourceLatLng != null && _destinationLatLng != null) {
-      Navigator.of(context).push(MaterialPageRoute(
-        builder: (context) => ConfirmRideScreen(
-          startTime: widget.startTime,
-          date: widget.date,
-          sourceAddress: _startsearchFieldController.text,
-          destinationAddress: _endsearchFieldController.text,
-          numSeats: widget.numSeats,
-          sourceLatLng: _sourceLatLng!,  // Use the actual source coordinates
-          destinationLatLng: _destinationLatLng!,  // Use the actual destination coordinates
-        ),
-      ));
-    }
-  }
-
-
-
 
   @override
   void initState() {
@@ -83,6 +63,7 @@ class _AddRideMapScreenState extends State<AddRideMapScreen> {
                 initialCameraPosition: initialPosition,
                 zoomControlsEnabled: false,
                 myLocationEnabled: true,
+                myLocationButtonEnabled: false,
                 markers: markers,
                 mapType: MapType.normal,
                 onMapCreated: (GoogleMapController controller) {
@@ -90,124 +71,145 @@ class _AddRideMapScreenState extends State<AddRideMapScreen> {
                   _determineLocation();
                 }
             ),
-            Positioned(
-                bottom: 2,
-                left: 0,
-                right: 0,
-                child: Padding(
-                  padding: EdgeInsets.all(10),
-                  child: Container(
-                    color: Colors.white,
-                    padding: EdgeInsets.fromLTRB(15, 10, 15, 10),
-                    child: Column(
-                      children: [
-                        Row(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Padding(
-                              padding: const EdgeInsets.fromLTRB(10, 5, 0, 0),
-                              child: Text("I am headed from..",
-                                textAlign: TextAlign.left,
-                                style: TextStyle(
-                                  fontSize: 30,
-                                  fontFamily: 'DMSans',
-                                  fontWeight: FontWeight.bold,
+            Form(
+              key: _formKey,
+              child: Positioned(
+                  bottom: 2,
+                  left: 0,
+                  right: 0,
+                  child: Padding(
+                    padding: EdgeInsets.all(10),
+                    child: Container(
+                      color: Colors.white,
+                      padding: EdgeInsets.fromLTRB(15, 10, 15, 10),
+                      child: Column(
+                        children: [
+                          Row(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Padding(
+                                padding: const EdgeInsets.fromLTRB(10, 5, 0, 0),
+                                child: Text("I am headed from..",
+                                  textAlign: TextAlign.left,
+                                  style: TextStyle(
+                                    fontSize: 30,
+                                    fontFamily: 'DMSans',
+                                    fontWeight: FontWeight.bold,
+                                  ),
                                 ),
                               ),
-                            ),
-                          ],
-                        ),
-                        SizedBox(height: 10),
-                        Row(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Container(
-                              width: 250,
-                              child: Padding(
-                                padding: const EdgeInsets.fromLTRB(10, 0, 0, 0),
-                                child: Divider(
-                                  color: Colors.grey[200],
-                                  thickness: 4, // Adjust the thickness as needed
+                            ],
+                          ),
+                          SizedBox(height: 10),
+                          Row(
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Container(
+                                width: 250,
+                                child: Padding(
+                                  padding: const EdgeInsets.fromLTRB(10, 0, 0, 0),
+                                  child: Divider(
+                                    color: Colors.grey[200],
+                                    thickness: 4, // Adjust the thickness as needed
+                                  ),
                                 ),
                               ),
-                            ),
-                          ],
-                        ),
-                        SizedBox(height: 20),
-                        TextField(
-                          controller: _startsearchFieldController,
-                          autofocus: false,
-                          decoration: InputDecoration(
-                            contentPadding: EdgeInsets.symmetric(vertical: 5.0, horizontal: 15.0),
-                            labelText: 'Start Location',
-                            floatingLabelBehavior: FloatingLabelBehavior.auto,
-                            filled: true,
-                            fillColor: Colors.grey[100],
-                            border: OutlineInputBorder(
-                              borderRadius: BorderRadius.circular(10.0),
-                              borderSide: BorderSide.none,
-                            ),
-                            prefixIcon: Icon(Icons.radio_button_checked, color: Colors.deepOrange[100]),
+                            ],
                           ),
-                          readOnly: true,
-                          onTap: () => _openAutoComplete(context, _startsearchFieldController, true),
-                        ),
-                        SizedBox(height: 15),
-                        TextField(
-                          controller: _endsearchFieldController,
-                          autofocus: false,
-                          decoration: InputDecoration(
-                            contentPadding: EdgeInsets.symmetric(vertical: 5.0, horizontal: 15.0),
-                            labelText: 'Destination Location',
-                            floatingLabelBehavior: FloatingLabelBehavior.auto,
-                            filled: true,
-                            fillColor: Colors.grey[100],
-                            border: OutlineInputBorder(
-                              borderRadius: BorderRadius.circular(10.0),
-                              borderSide: BorderSide.none,
+                          SizedBox(height: 20),
+                          TextFormField(
+                            controller: _startsearchFieldController,
+                            autofocus: false,
+                            decoration: InputDecoration(
+                              contentPadding: EdgeInsets.symmetric(vertical: 5.0, horizontal: 15.0),
+                              labelText: 'Start Location',
+                              floatingLabelBehavior: FloatingLabelBehavior.auto,
+                              filled: true,
+                              fillColor: Colors.grey[100],
+                              border: OutlineInputBorder(
+                                borderRadius: BorderRadius.circular(10.0),
+                                borderSide: BorderSide.none,
+                              ),
+                              prefixIcon: Icon(Icons.radio_button_checked, color: Colors.deepOrange[100]),
                             ),
-                            prefixIcon: Icon(Icons.radio_button_checked, color: Colors.deepPurple[100]),
-                          ),
-                          readOnly: true,
-                          onTap: () => _openAutoComplete(context, _endsearchFieldController, false),
-                        ),
-                        SizedBox(height: 20),
-                        Container(
-                          height: 50,
-                          width: 350,
-                          child: ElevatedButton(
-                            onPressed: () {
-                              print("Start Text: ${_startsearchFieldController.text}");
-                              print("End Text: ${_endsearchFieldController.text}");
-                              print("Source LatLng: $_sourceLatLng");
-                              print("Destination LatLng: $_destinationLatLng");
-                              if (_startsearchFieldController.text.isNotEmpty &&
-                                  _endsearchFieldController.text.isNotEmpty &&
-                                  _sourceLatLng != null &&
-                                  _destinationLatLng != null) {
-                                _goToConfirmationScreen(context);
+                            readOnly: true,
+                            validator: (value) {
+                              if (value == null || value.isEmpty) {
+                                return 'Please enter start location';
                               }
+                              return null;
                             },
-                            child: Text("Submit ride", style: TextStyle(
-                              fontSize: 17,
-                              fontFamily: 'DMSans',
-                              fontWeight: FontWeight.normal,
-                            )
+                            onTap: () => _openAutoComplete(context, _startsearchFieldController, true),
+                          ),
+                          SizedBox(height: 15),
+                          TextFormField(
+                            controller: _endsearchFieldController,
+                            autofocus: false,
+                            decoration: InputDecoration(
+                              contentPadding: EdgeInsets.symmetric(vertical: 5.0, horizontal: 15.0),
+                              labelText: 'Destination Location',
+                              floatingLabelBehavior: FloatingLabelBehavior.auto,
+                              filled: true,
+                              fillColor: Colors.grey[100],
+                              border: OutlineInputBorder(
+                                borderRadius: BorderRadius.circular(10.0),
+                                borderSide: BorderSide.none,
+                              ),
+                              prefixIcon: Icon(Icons.radio_button_checked, color: Colors.deepPurple[100]),
                             ),
-                            style: ElevatedButton.styleFrom(
-                                shape: RoundedRectangleBorder(
-                                  borderRadius: BorderRadius.circular(6), // Adjust the border radius here
-                                ),
-                                foregroundColor: Colors.white, // Change the background color here
-                                backgroundColor: Colors.black38, // Change the text color here
-                                padding: EdgeInsets.fromLTRB(0, 15, 0, 15)
+                            readOnly: true,
+                            validator: (value) {
+                              if (value == null || value.isEmpty) {
+                                return 'Please enter destination location';
+                              }
+                              return null;
+                            },
+                            onTap: () => _openAutoComplete(context, _endsearchFieldController, false),
+                          ),
+                          SizedBox(height: 20),
+                          Container(
+                            height: 50,
+                            width: 350,
+                            child: ElevatedButton(
+                              onPressed: () {
+                              if (_formKey.currentState!.validate()) {
+                              Navigator.of(context).push(MaterialPageRoute(builder: (BuildContext context){
+                                return ConfirmRideScreen(
+                                  startTime: widget.startTime,
+                                  date: widget.date,
+                                  sourceAddress: _startsearchFieldController.text,
+                                  destinationAddress: _endsearchFieldController.text,
+                                  numSeats: widget.numSeats,
+                                  sourceLatLng: _sourceLatLng!,  // Use the actual source coordinates
+                                  destinationLatLng: _destinationLatLng!,  // Use the actual destination coordinates
+                                );
+                              }));
+                              }
+                              else{
+                              // do noting when form validation fails => stays on same screen
+                              }
+                              },
+                              child: Text("Submit ride", style: TextStyle(
+                                fontSize: 17,
+                                fontFamily: 'DMSans',
+                                fontWeight: FontWeight.normal,
+                              )
+                              ),
+                              style: ElevatedButton.styleFrom(
+                                  shape: RoundedRectangleBorder(
+                                    borderRadius: BorderRadius.circular(6), // Adjust the border radius here
+                                  ),
+                                  foregroundColor: Colors.white, // Change the background color here
+                                  backgroundColor: Colors.black38, // Change the text color here
+                                  padding: EdgeInsets.fromLTRB(0, 15, 0, 15)
+                              ),
                             ),
                           ),
-                        ),
-                      ],
+                        ],
+                      ),
                     ),
-                  ),
-                )
+                  )
+              ),
             ),
           ],
         )
